requirements.txt
# Runtime (API) — graph stack (langgraph, langchain-core) pinned in Story 3.x
fastapi==0.115.8
uvicorn[standard]==0.34.0

# Persistence (Story 1.2)
sqlalchemy==2.0.40
alembic==1.15.2
asyncpg==0.30.0
psycopg[binary]==3.2.6

# Auth (Story 1.3)
argon2-cffi==23.1.0
PyJWT==2.10.1
python-multipart==0.0.20
email-validator==2.2.0

# Dev / test (single-file bootstrap; split later if you add a prod-only constraints file)
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.3
